\name{many_sample_means}
\alias{many_sample_means}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Takes reps many sample means
}
\usage{
sample_Mean(vec, n, reps)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{
  A vector which you want to take the mean of
  }
  \item{n}{
  Sample Size
  }
  \item{reps}{
  Amount of sample means
  }
}
\value{
A vector of reps sample means
}
