\name{sample_Mean}
\alias{sample_Mean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Takes a sample from vector and returns a mean
}
\usage{
sample_Mean(vec, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{
  A vector which you want to take the mean of
  }
  \item{n}{
  Sample Size
  }
}
\value{
A double which is the mean of the sample
}
