\name{sample_means_ns}
\alias{sample_means_ns}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Takes reps many sample means for ns sample sizes
}
\usage{
sample_Mean(vec, ns, reps)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{
  A vector which you want to take the mean of
  }
  \item{ns}{
  Veector of Sample Size
  }
  \item{reps}{
  Amount of sample means
  }
}
\value{
A data frame with Sample means and ns
}
